Initial release

+++ app/Http/Controllers/ParticipantApi/FileController.php
@@ -2,26 +2,29 @@
 
 namespace BristolSU\Module\UploadFile\Http\Controllers\ParticipantApi;
 
+use BristolSU\Module\UploadFile\Events\DocumentDeleted;
+use BristolSU\Module\UploadFile\Events\DocumentUpdated;
 use BristolSU\Module\UploadFile\Events\DocumentUploaded;
 use BristolSU\Module\UploadFile\Http\Controllers\Controller;
 use BristolSU\Module\UploadFile\Http\Requests\ParticipantApi\FileController\StoreRequest;
 use BristolSU\Module\UploadFile\Models\File;
 use BristolSU\Support\Activity\Activity;
+use BristolSU\Support\ActivityInstance\Contracts\ActivityInstanceResolver;
 use BristolSU\Support\Authentication\Contracts\Authentication;
 use BristolSU\Support\ModuleInstance\ModuleInstance;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\Response;
 
 class FileController extends Controller
 {
     public function store(StoreRequest $request, Authentication $authentication)
     {
         $fileMetadata = collect();
-        
+
         foreach (Arr::wrap($request->file('file')) as $file) {
 
-            $path = $file->store(alias());
+            $path = $file->store('uploadfile');
 
             $fileMetadata->push($tempFileMeta = File::create([
                 'title' => $request->input('title'),
@@ -32,7 +35,7 @@ class FileController extends Controller
                 'size' => $file->getSize(),
                 'uploaded_by' => $authentication->getUser()->id(),
             ]));
-            
+
             event(new DocumentUploaded($tempFileMeta));
         }
 
@@ -42,21 +45,32 @@ class FileController extends Controller
     public function index(Request $request)
     {
         $this->authorize('file.index');
-        return File::forResource()->with('statuses')->get();
+        return File::forResource()->with(['statuses', 'comments'])->get();
     }
 
     public function destroy(Request $request, Activity $activity, ModuleInstance $moduleInstance, File $file)
     {
         $this->authorize('file.destroy');
-        
+
+        if((int) $file->activity_instance_id !== (int) app(ActivityInstanceResolver::class)->getActivityInstance()->id) {
+            throw new AuthorizationException();
+        }
+                
         $file->delete();
+        $file->refresh();
+
+        event(new DocumentDeleted($file));
         
-        return $file;
+        return $file->refresh();
     }
 
     public function show(Request $request, Activity $activity, ModuleInstance $moduleInstance, File $file)
     {
         $this->authorize('file.index');
+
+        if((int) $file->activity_instance_id !== (int) app(ActivityInstanceResolver::class)->getActivityInstance()->id) {
+            throw new AuthorizationException();
+        }
         
         return $file->load('statuses');
     }
@@ -65,13 +79,18 @@ class FileController extends Controller
     {
         $this->authorize('file.update');
 
+        if((int) $file->activity_instance_id !== (int) app(ActivityInstanceResolver::class)->getActivityInstance()->id) {
+            throw new AuthorizationException();
+        }
+        
         $file->title = $request->input('title', $file->title);
         $file->description = $request->input('description', $file->description);
+
+        $file->save();
+            
+        event(new DocumentUpdated($file));
         
-        if($file->save()) {
-            return $file;
-        }
-        return Response::make('Could not save the file', 500);
+        return $file;
     }
 
 }
